Updated the quantization script.
