+++ model_bf16_to_fp16_converter/data_axs.json
@@ -5,7 +5,7 @@
     ],
 
     "pipeline": [ "^^", "execute", [[
-        [ "convert" ],
+        [ "run" ],
         [ ],
         [ "get", "stored_newborn_entry" ]
     ]] ],
