+++ quantize_llm_compressor_recipe/data_axs.json
@@ -1,11 +1,11 @@
 {
-    "_parent_entries": [ [ "^", "byname", "entry_creator" ] ],
+    "_parent_entries": [ [ "^", "byname", "python_script" ], [ "^", "byname", "entry_creator" ] ],
     "_producer_rules": [
         [ [ "quantized", "method=llm_compressor" ], [[ "get", "pipeline" ]] ]
     ],
 
     "pipeline": [ "^^", "execute", [[
-        [ "quantize" ],
+        [ "run" ],
         [ ],
         [ "get", "stored_newborn_entry" ]
     ]] ], 
@@ -52,54 +52,25 @@
     "max_sequence_length": 4096,
     "num_calibration_samples": 1000,
 
-    "transformers_query": [
-        "python_package",
-        "package_name=transformers",
-        "package_version=4.43.2",
-        ["desired_python_version", [ "^^", "get", "desired_python_version" ] ]
+    "python_deps": [
+        [ "^^", "python_sync_pip_package", [[ 
+            "python_package",
+            "package_name=llm-compressor",
+            "installable=git+https://github.com/vllm-project/llm-compressor.git@sa/big_model_support",
+            ["desired_python_version", [ "^^", "get", "desired_python_version" ] ]
+        ]] ]
     ],
-    "torch_query":  [
-        "python_package",
-        "package_name=torch",
-        "package_version=2.4.0", 
-        ["desired_python_version", [ "^^", "get", "desired_python_version" ] ]
-    ],
-    "llm_compressor_query":  [
-        "python_package",
-        "package_name=llm-compressor",
-        "installable=git+https://github.com/vllm-project/llm-compressor.git@sa/big_model_support",
-        ["desired_python_version", [ "^^", "get", "desired_python_version" ] ]
-    ],
-    "compressed_tensors_query":  [
-        "python_package",
-        "package_name=compressed-tensors",
-        "package_version=0.5.0",
-        ["desired_python_version", [ "^^", "get", "desired_python_version" ] ]
-    ],
-
-    "_BEFORE_CODE_LOADING": [ "^^", "execute", [[
-        ["get_kernel"],
-        ["byquery", [[ "^^", "get", "transformers_query" ]] ],
-        ["use"],
-        [],
-        ["get_kernel"],
-        ["byquery", [[ "^^", "get", "torch_query" ]] ],
-        ["use"],
-        [],
-        ["get_kernel"],
-        ["byquery", [[ "^^", "get", "llm_compressor_query" ]] ],
-        ["use"],
-        [],
-        ["get_kernel"],
-        ["byquery", [[ "^^", "get", "compressed_tensors_query" ]] ],
-        ["use"],
-        []
-    ]] ],
 
     "newborn_entry_tags": [ "quantized", "method=llm_compressor" ],
     "newborn_name_template": [ "quantized_#{model_name}#_#{model_variant}#_using_#{method}#" ],
     "newborn_entry_param_names": [
         "model_name",
         "model_variant"
+    ],
+
+    "rel_script_path": "quantize.py",
+
+    "script_extra_params": [ "^^", "substitute", 
+        "#{source_model_path}# #{target_model_path}# #{calib_dataset_path}# #{num_gpus}# #{max_sequence_length}# #{num_calibration_samples}#"
     ]
 }
